+++ CHANGELOG.md
@@ -13,6 +13,7 @@ AdminNeo 5.5.0
 - Display data length and index length for materialized views (by @vrana)
 - Link routines from syntax highlighting (by @vrana)
 - Do not quote 0 in CSV export (by @vrana)
+- Add getServerVariables() and getStatusVariables() customization methods
 - PostgreSQL: Add SQL operator to select (by @vrana)
 - PostgreSQL: Display index expressions (by @vrana)
 - PostgreSQL: Shorten values in hstore columns (by @vrana)